Error test for list_capteur
