+++ SERVER/routes/sample.js
@@ -1,18 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const db = require('../database');
-
-const UNITS = {
-    temperature: "C",
-    pressure: "hP",
-    humidity: "%",
-    rain: "mm/m2",
-    lux: "Lux",
-    wind_heading: "°",
-    wind_speed_avg: "km/h",
-    lat: "DD",
-    lon: "DD"
-};
+const { UNITS, CAPTEURS }= require('../data.js');
 
 // Route /sample/:start/:stop
 router.get('/:start/:stop', async (req, res) => {
@@ -37,7 +26,17 @@ router.get('/:start/:stop', async (req, res) => {
 router.get('/:start/:stop/:list_capteur', async (req, res) => {
     try {
         const { start, stop, list_capteur } = req.params;
-        const capteurs = list_capteur.split("-"); // Transformer en tableau
+        
+        if (!list_capteur.includes('-')) {
+            return res.status(400).json({ message: "A query argument is invalid" });
+        }
+
+        const capteurs = list_capteur.split("-");
+
+        const capteursInvalides = capteurs.filter(capteur => !VALID_CAPTEURS.includes(capteur));
+        if (capteursInvalides.length > 0) {
+            return res.status(400).json({ message: "A query argument is invalid"});
+        }
 
         const sampleData = await db.getSampleDataBySensor(start, stop, capteurs);
